fix type miscast

+++ SonicSole.cpp
@@ -125,7 +125,7 @@ void SonicSole::toCSV(double time, double heelpresh, double forepresh, float az)
 }
 
 double SonicSole::getRunningTime() {
-    return (currentTime - startTime) / 1000000.0;
+    return ((double) currentTime - (double) startTime) / 1000000.0;
 }
 
 void SonicSole::updateCurrentTime() {
